Refactor : cleansing of all codes, add README.md file, seed bug solved.
