+++ papers/QTRL/configs/defaults.json
@@ -1,0 +1,11 @@
+{
+  "description" : "Default parameters for the Reinforcement Learning Train (REINFORCE)",
+  "env_name": "CartPole",
+  "q_output_size": 4,
+  "nb_photons": 2,
+  "nb_modes": 3,
+  "hidden_sizes": [8, 8],
+  "final_output_size": 8,
+  "num_episodes": 500,
+  "lr": 0.001
+}
