+++ papers/QTRL/requirements.txt
@@ -1,256 +1,5 @@
-absl-py==2.4.0
-alabaster==1.0.0
-alembic==1.18.4
-annotated-types==0.7.0
-anyio==4.13.0
-appdirs==1.4.4
-argon2-cffi==25.1.0
-argon2-cffi-bindings==25.1.0
-arrow==1.4.0
-asttokens==3.0.1
-astunparse==1.6.3
-async-lru==2.3.0
-attrs==26.1.0
-autograd==1.8.0
-autoray==0.8.2
-babel==2.18.0
-beautifulsoup4==4.14.3
-bleach==6.3.0
-cachetools==7.0.6
-certifi==2026.2.25
-cffi==2.0.0
-charset-normalizer==3.4.7
-cloudpickle==3.1.2
-colorlog==6.10.1
-comm==0.2.3
-commonmark==0.9.1
-contourpy==1.3.3
-cryptography==46.0.7
-cuda-bindings==13.2.0
-cuda-pathfinder==1.5.3
-cuda-toolkit==13.0.2
-cycler==0.12.1
-debugpy==1.8.20
-decorator==5.2.1
-defusedxml==0.7.1
-diastatic-malt==2.15.2
-dill==0.3.4
-docutils==0.22.4
-drawsvg==2.4.1
-executing==2.2.1
-exqalibur==1.1.1
-Farama-Notifications==0.0.4
-fastjsonschema==2.21.2
-filelock==3.28.0
-flatbuffers==25.12.19
-fonttools==4.62.1
-fqdn==1.5.1
-fsspec==2026.3.0
-gast==0.7.0
-google-pasta==0.2.0
-greenlet==3.5.0
-grpcio==1.80.0
-gym==0.26.2
-gym-notices==0.1.0
-gymnasium==1.0.0
-h11==0.16.0
-h5py==3.14.0
-httpcore==1.0.9
-httpx==0.28.1
-ibm-cloud-sdk-core==3.24.4
-ibm-platform-services==0.75.0
-ibm-quantum-schemas==0.5.20260320
-idna==3.11
-imagesize==2.0.0
-ipykernel==7.2.0
-ipython==9.12.0
-ipython_pygments_lexers==1.1.1
-ipywidgets==8.1.8
-isoduration==20.11.0
-jedi==0.19.2
-Jinja2==3.1.6
-joblib==1.5.3
-json5==0.14.0
-jsonpointer==3.1.1
-jsonschema==4.26.0
-jsonschema-specifications==2025.9.1
-jupyter==1.1.1
-jupyter-console==6.6.3
-jupyter-events==0.12.0
-jupyter-lsp==2.3.1
-jupyter_client==8.8.0
-jupyter_core==5.9.1
-jupyter_server==2.17.0
-jupyter_server_terminals==0.5.4
-jupyterlab==4.5.6
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.28.0
-jupyterlab_widgets==3.0.16
-keras==3.14.0
-kiwisolver==1.5.0
-lark==1.3.1
-latexcodec==3.0.1
-libclang==18.1.1
-loguru==0.7.3
-Mako==1.3.12
-Markdown==3.10.2
-markdown-it-py==4.0.0
-MarkupSafe==3.0.3
-matplotlib==3.10.8
-matplotlib-inline==0.2.1
-mdurl==0.1.2
 merlinquantum==0.2.3
-mistune==3.2.0
-ml_dtypes==0.5.4
-mpmath==1.3.0
-msgpack==1.1.2
-msgpack-numpy==0.4.8
-multimethod==2.0.2
-multipledispatch==1.0.0
-multiprocess==0.70.12.2
-namex==0.1.0
-nbclient==0.10.4
-nbconvert==7.17.1
-nbformat==5.10.4
-nbsphinx==0.9.8
-nest-asyncio==1.6.0
-networkx==3.6.1
-notebook==7.5.5
-notebook_shim==0.2.4
-numpy==2.4.4
-nvidia-cublas==13.1.0.3
-nvidia-cublas-cu12==12.8.4.1
-nvidia-cuda-cupti==13.0.85
-nvidia-cuda-cupti-cu12==12.8.90
-nvidia-cuda-nvrtc==13.0.88
-nvidia-cuda-nvrtc-cu12==12.8.93
-nvidia-cuda-runtime==13.0.96
-nvidia-cuda-runtime-cu12==12.8.90
-nvidia-cudnn-cu12==9.10.2.21
-nvidia-cudnn-cu13==9.19.0.56
-nvidia-cufft==12.0.0.61
-nvidia-cufft-cu12==11.3.3.83
-nvidia-cufile==1.15.1.6
-nvidia-cufile-cu12==1.13.1.3
-nvidia-curand==10.4.0.35
-nvidia-curand-cu12==10.3.9.90
-nvidia-cusolver==12.0.4.66
-nvidia-cusolver-cu12==11.7.3.90
-nvidia-cusparse==12.6.3.3
-nvidia-cusparse-cu12==12.5.8.93
-nvidia-cusparselt-cu12==0.7.1
-nvidia-cusparselt-cu13==0.8.0
-nvidia-nccl-cu12==2.27.3
-nvidia-nccl-cu13==2.28.9
-nvidia-nvjitlink==13.0.88
-nvidia-nvjitlink-cu12==12.8.93
-nvidia-nvshmem-cu13==3.4.5
-nvidia-nvtx==13.0.85
-nvidia-nvtx-cu12==12.8.90
-opt_einsum==3.4.0
-optree==0.19.0
-optuna==4.8.0
-orjson==3.11.8
-packaging==26.0
-pandas==3.0.2
-pandocfilters==1.5.1
-parso==0.8.6
-pathos==0.2.8
-pennylane==0.44.1
-pennylane_lightning==0.44.0
-perceval-quandela==1.1.0
-pexpect==4.9.0
-pillow==12.2.0
-platformdirs==4.9.4
-ply==3.11
-pox==0.3.7
-ppft==1.7.8
-prometheus_client==0.24.1
-prompt_toolkit==3.0.52
-protobuf==7.34.1
-psutil==7.2.2
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pybase64==1.4.3
-pycparser==3.0
-pydantic==2.12.5
-pydantic_core==2.41.5
-pygame==2.6.1
-Pygments==2.20.0
-PyJWT==2.12.1
-pylatexenc==2.10
-pyparsing==3.3.2
-pyscf==2.13.0
-pyspnego==0.12.1
-python-dateutil==2.9.0.post0
-python-json-logger==4.1.0
-PyYAML==6.0.3
-pyzmq==27.1.0
-qiskit==2.4.0
-qiskit-ibm-runtime==0.46.1
-qiskit-terra==0.46.3
-recommonmark==0.7.1
-referencing==0.37.0
-requests==2.33.1
-requests_ntlm==1.3.0
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rfc3987-syntax==1.1.0
-rich==15.0.0
-roman-numerals==4.1.0
-rpds-py==0.30.0
-rustworkx==0.17.1
-samplomatic==0.17.1
-scikit-learn==1.8.0
-scipy==1.17.1
-scipy-openblas32==0.3.31.188.0
-Send2Trash==2.1.0
-setuptools==81.0.0
-six==1.17.0
-snowballstemmer==3.0.1
-soupsieve==2.8.3
-Sphinx==9.1.0
-sphinxcontrib-applehelp==2.0.0
-sphinxcontrib-devhelp==2.0.0
-sphinxcontrib-htmlhelp==2.1.0
-sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==2.0.0
-sphinxcontrib-serializinghtml==2.0.0
-SQLAlchemy==2.0.49
-stack-data==0.6.3
-stevedore==5.7.0
-symengine==0.14.1
-sympy==1.14.0
-tabulate==0.10.0
-tensorboard==2.20.0
-tensorboard-data-server==0.7.2
-tensorflow==2.21.0
-tensorpack==0.11
-termcolor==3.3.0
-terminado==0.18.1
-threadpoolctl==3.6.0
-tinycss2==1.4.0
-toml==0.10.2
-tomlkit==0.14.0
-torch==2.11.0
-torchdiffeq==0.2.5
-torchpack==0.3.1
-torchquantum==0.1.5
-torchvision==0.26.0
-tornado==6.5.5
-tqdm==4.67.3
-traitlets==5.14.3
-triton==3.6.0
-typing-inspection==0.4.2
-typing_extensions==4.15.0
-tzdata==2026.1
-uri-template==1.3.0
-urllib3==2.6.3
-wcwidth==0.6.0
-webcolors==25.10.0
-webencodings==0.5.1
-websocket-client==1.9.0
-Werkzeug==3.1.8
-wheel==0.47.0
-widgetsnbextension==4.0.15
-wrapt==2.1.2
+torch
+numpy
+gymnasium[classic_control]
+minigrid
